Did some refactoring of using self instead of "_"
//Visually
added background to the controller also added a uiimage to highlight behind the players image(for who ever turn it is their image view will be on).
//ViewController
added a timer to update health ever x seconds.  Put it in the view controller view did load
also changed the name of method setScreeForPlayer to setUniqueLookForCurrentPlayerTurn and put it in the set screen method and used set screen to set up visual updates

//SpecialScreenViewController
updated the setSpecialScreen method to do more

//misc
added some more images